Add support for setupTestFrameworkScriptFile option.
Jest allows for an option called `setupTestFrameworkScriptFile` that is defined
as a path to a module that is run before each of your test files. This commit
adds similar support to jest-runner-eslint.

This implementation will only actually run the setupTestFrameworkScriptFile one
time, since this runner executes the tests in persistent workers instead of
isolated vms.